Rename and rework entry point
