+++ Emby.Server.Implementations/EnvironmentInfo/EnvironmentInfo.cs
@@ -1,12 +1,12 @@
 ﻿using System;
 using System.IO;
 using MediaBrowser.Model.System;
+using System.Runtime.InteropServices;
 
 namespace Emby.Server.Implementations.EnvironmentInfo
 {
     public class EnvironmentInfo : IEnvironmentInfo
     {
-        private Architecture? _customArchitecture;
         private MediaBrowser.Model.System.OperatingSystem? _customOperatingSystem;
 
         public virtual MediaBrowser.Model.System.OperatingSystem OperatingSystem
@@ -60,22 +60,7 @@ namespace Emby.Server.Implementations.EnvironmentInfo
             }
         }
 
-        public Architecture SystemArchitecture
-        {
-            get
-            {
-                if (_customArchitecture.HasValue)
-                {
-                    return _customArchitecture.Value;
-                }
-
-                return Environment.Is64BitOperatingSystem ? MediaBrowser.Model.System.Architecture.X64 : MediaBrowser.Model.System.Architecture.X86;
-            }
-            set
-            {
-                _customArchitecture = value;
-            }
-        }
+        public Architecture SystemArchitecture { get; set; }
 
         public string GetEnvironmentVariable(string name)
         {
