+++ Jellyfin.Server/PowerManagement.cs
@@ -1,21 +1,23 @@
-﻿using System;
+using System;
 using MediaBrowser.Model.System;
 
-namespace MediaBrowser.Server.Mono.Native
+namespace Jellyfin.Server.Native
 {
     public class PowerManagement : IPowerManagement
     {
         public void PreventSystemStandby()
         {
+
         }
 
         public void AllowSystemStandby()
         {
+
         }
 
         public void ScheduleWake(DateTime wakeTimeUtc, string displayName)
         {
-            // nothing to Do
+
         }
     }
 }
