+++ Jellyfin.Server/SocketSharp/SharpWebSocket.cs
@@ -1,12 +1,12 @@
-﻿using MediaBrowser.Common.Events;
-using Microsoft.Extensions.Logging;
-using System;
+﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Net.WebSockets;
 using Emby.Server.Implementations.Net;
+using MediaBrowser.Common.Events;
+using Microsoft.Extensions.Logging;
 
-namespace EmbyServer.SocketSharp
+namespace Jellyfin.SocketSharp
 {
     public class SharpWebSocket : IWebSocket
     {
