--- MediaBrowser.Model/System/Architecture.cs
@@ -1,10 +1,0 @@
-﻿namespace MediaBrowser.Model.System
-{
-    public enum Architecture
-    {
-        X86 = 0,
-        X64 = 1,
-        Arm = 2,
-        Arm64 = 3
-    }
-}
