--- MediaBrowser.Server.Mono/Properties/launchSettings.json
@@ -1,7 +1,0 @@
-{
-  "profiles": {
-    "EmbyServer": {
-      "commandName": "Project"
-    }
-  }
-}
