--- MediaBrowser.Server.Mono/Resources/Configuration/logging.json
@@ -1,19 +1,0 @@
-{
-    "Serilog": {
-        "MinimumLevel": "Information",
-        "WriteTo": [
-            { "Name": "Console",
-                "Args": {
-                    "outputTemplate": "[{Timestamp:HH:mm:ss}] [{Level:u3}] {Message:lj}{NewLine}{Exception}"
-                }
-            },
-            { "Name": "File",
-                "Args": {
-                    "path": "%JELLYFIN_LOG_DIR%//log_.log",
-                    "rollingInterval": "Day",
-                    "outputTemplate": "[{Timestamp:yyyy-MM-dd HH:mm:ss.fff zzz}] [{Level:u3}] {Message}{NewLine}{Exception}"
-                }
-            }
-        ]
-    }
-}
